Rework the tests
